requirements.txt
openai
streamlit
python-dotenv
pandas
python-docx
pymupdf
konlpy
openpyxl
